requirements.txt
Flask
google-play-scraper
transformers
torch
keybert
scikit-learn
google-generativeai
gunicorn